requirements.txt
django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
markdown==3.6
numpy==2.0.1
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1